% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{sample_data_extern}
\alias{sample_data_extern}
\title{Sample External Data}
\format{
A data frame with 49 rows and 51 columns (1 response variable and 50 predictors):
\describe{
  \item{y}{Response variable (factor)}
  \item{x1}{First predictor variable (numeric)}
  \item{x2}{Second predictor variable (numeric)}
  ...
  \item{x50}{Fiftieth predictor variable (numeric)}
}
}
\source{
Generated from `data10` in the package development process.
}
\usage{
sample_data_extern
}
\description{
A dataset containing the response variable `y` and 50 selected predictor variables from `data10`.
}
\keyword{datasets}
