% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuneandtrainRobustTuneCLasso}
\alias{tuneandtrainRobustTuneCLasso}
\title{Tune and Train RobustTuneC Lasso}
\usage{
tuneandtrainRobustTuneCLasso(data, dataext, maxit = 120000, nlambda = 100)
}
\arguments{
\item{data}{Training data as a data frame. The first column should be the response variable.}

\item{dataext}{External validation data as a data frame. The first column should be the response variable.}

\item{maxit}{Maximum number of iterations. Default is 120000.}

\item{nlambda}{The number of lambda values to use for cross-validation.}
}
\value{
A list containing the best lambda (`best_lambda`) and the final trained model (`best_model`).
}
\description{
This function tunes and trains a Lasso classifier with "RobustTuneC" method, using 5-fold cross-validation 
and selects the best model based on AUC.
}
