% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuneandtrainRobustTuneCRF}
\alias{tuneandtrainRobustTuneCRF}
\title{Tune and Train RobustTuneC Random Forest}
\usage{
tuneandtrainRobustTuneCRF(data, dataext, num.trees = 500)
}
\arguments{
\item{data}{Training data as a data frame. The first column should be the response variable.}

\item{dataext}{External validation data as a data frame. The first column should be the response variable.}

\item{num.trees}{Number of trees to grow. Default is 500.}

\item{maxit}{Maximum number of iterations. Default is 120000.}
}
\value{
A list containing the best minimum node size (`best_min_node_size`) and the final trained model (`best_model`).
}
\description{
This function tunes and trains a Random Forest classifier with "RobustTuneC" method, using 5-fold cross-validation 
and selects the best model based on AUC.
}
