% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuneandtrainRobustTuneC}
\alias{tuneandtrainRobustTuneC}
\title{Tune and Train by tuning method RobustTuneC}
\usage{
tuneandtrainRobustTuneC(data, dataext, classifier, ...)
}
\arguments{
\item{data}{Training data as a data frame. The first column should be the response variable.}

\item{dataext}{External validation data as a data frame. The first column should be the response variable.}

\item{classifier}{The classifier to use. Must be one of 'boosting', 'rf', 'lasso', 'ridge', 'svm'.}

\item{...}{Additional arguments to pass to the specific classifier function.}
}
\value{
A list containing the results from the specific classifier tuning and training function.
}
\description{
This function tunes and trains a specified classifier using the appropriate method and data.
}
