% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuneandtrainRobustTuneCBoost}
\alias{tuneandtrainRobustTuneCBoost}
\title{Tune and Train RobustTuneC Boosting}
\usage{
tuneandtrainRobustTuneCBoost(
  data,
  dataext,
  mstop_seq = seq(5, 1000, by = 5),
  nu = 0.1
)
}
\arguments{
\item{data}{Training data as a data frame. The first column should be the response variable.}

\item{dataext}{External validation data as a data frame. The first column should be the response variable.}

\item{mstop_seq}{A sequence of boosting iterations to consider. Default is a sequence starting at 5 and increasing by 5 each time, up to 1000.}

\item{nu}{Learning rate for the boosting algorithm. Default is 0.1.}
}
\value{
A list containing the best number of boosting iterations (`best_mstop`) and the final trained model (`best_model`).
}
\description{
This function tunes and trains a Boosting classifier with "RobustTuneC" method, using 5-fold cross-validation 
and selects the best model based on AUC.
}
\examples{
\dontrun{
# Load the sample data
data(sample_data_train)
data(sample_data_extern)

# Example usage with the sample data
mstop_seq <- seq(50, 500, by = 50)
result <- tuneandtrainRobustTuneCBoost(sample_data_train, sample_data_extern, mstop_seq)
result$best_mstop
result$best_model
}
}
