% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneandtrainRobustTuneCRF.R
\name{tuneandtrainRobustTuneCRF}
\alias{tuneandtrainRobustTuneCRF}
\title{Tune and Train RobustTuneC Random Forest}
\usage{
tuneandtrainRobustTuneCRF(data, dataext, K = 5, num.trees = 500)
}
\arguments{
\item{data}{A data frame containing the training data. The first column should be the response variable (factor), and the remaining columns should be the predictor variables.}

\item{dataext}{A data frame containing the external validation data. The first column should be the response variable (factor), and the remaining columns should be the predictor variables.}

\item{K}{Number of folds to use in cross-validation. Default is 5.}

\item{num.trees}{An integer specifying the number of trees to grow in the Random Forest. Default is 500.}
}
\value{
A list containing the best minimum node size (`best_min_node_size`), the final trained model (`best_model`), and the AUC of the final model (`final_auc`).
}
\description{
This function tunes and trains a Random Forest classifier using the "RobustTuneC" method. The function 
uses K-fold cross-validation (with K specified by the user) to select the best model based on AUC (Area Under the Curve).
}
\examples{
\dontrun{
# Load sample data
data(sample_data_train)
data(sample_data_extern)

# Example usage
result <- tuneandtrainRobustTuneCRF(sample_data_train, sample_data_extern, K = 5, num.trees = 500)
result$best_min_node_size
result$best_model
result$final_auc
}
}
