% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuneandtrainRobustTuneCSVM}
\alias{tuneandtrainRobustTuneCSVM}
\title{Tune and Train RobustTuneC Support Vector Machine(SVM)}
\usage{
tuneandtrainRobustTuneCSVM(
  data,
  dataext,
  seed = 123,
  kernel = "linear",
  cost_seq = 2^(-15:15),
  scale = FALSE
)
}
\arguments{
\item{data}{Training data as a data frame. The first column should be the response variable.}

\item{dataext}{External validation data as a data frame. The first column should be the response variable.}

\item{seed}{Random seed for reproducibility. Default is 123.}

\item{kernel}{The kernel type to be used in the algorithm. It can be "linear", "polynomial", "radial", or "sigmoid". Default is "linear".}

\item{cost_seq}{A sequence of cost values to consider for cross-validation. Default is `2^(-15:15)`.}

\item{scale}{A logical vector indicating the variables to be scaled. Default is `FALSE`.}
}
\value{
A list containing the best cost (`best_cost`) and the final trained model (`best_model`).
}
\description{
This function tunes and trains a SVM classifier with "RobustTuneC" method, using 5-fold cross-validation 
and selects the best model based on AUC.
}
